@@ -12,30 +12,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
+    "@angular/animations": "^9.1.9",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "9.1.9",
+    "@angular/compiler": "9.1.9",
+    "@angular/core": "9.1.9",
+    "@angular/forms": "9.1.9",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "9.1.9",
+    "@angular/platform-browser-dynamic": "9.1.9",
+    "@angular/platform-server": "9.1.9",
+    "@angular/router": "9.1.9",
+    "@fortawesome/fontawesome-free": "^5.13.0",
     "@nguniversal/module-map-ngfactory-loader": "8.1.1",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.3.1",
     "core-js": "^3.3.3",
+    "font-awesome": "^4.7.0",
+    "http-errors": "^1.7.3",
     "jquery": "3.4.1",
     "oidc-client": "^1.9.1",
     "popper.js": "^1.16.0",
     "rxjs": "^6.5.3",
-    "zone.js": "0.9.1"
+    "rxjs-compat": "^6.5.5",
+    "saturn-datepicker": "^8.0.5",
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.803.14",
-    "@angular/cli": "8.3.14",
-    "@angular/compiler-cli": "8.2.12",
-    "@angular/language-service": "8.2.12",
+    "@angular/cli": "^9.1.7",
+    "@angular/compiler-cli": "9.1.9",
+    "@angular/language-service": "9.1.9",
     "@types/jasmine": "~3.4.4",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "~12.11.6",
@@ -47,7 +54,7 @@
     "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3"
   },
   "optionalDependencies": {
     "node-sass": "^4.12.0",
