@@ -1,19 +1,16 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 
 namespace TaskAPI.Models
 {
-    public class TaskContext : DbContext
+    public class TaskContext : IdentityDbContext
     {
+        public DbSet<TaskItem> Tasks { get; set; }
+        public DbSet<Comment> Comments { get; set; }
+
         public TaskContext(DbContextOptions<TaskContext> options)
             : base(options)
         {
         }
-
-        public DbSet<TaskItem> Tasks { get; set; }
-        public DbSet<Comment> Comments { get; set; }
     }
 }
