@@ -3,27 +3,23 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using TaskAPI.ViewModel;
 
 namespace TaskAPI.Models.Validators
 {
-    public class TaskValidator : AbstractValidator<TaskItem>
+    public class TaskValidator : AbstractValidator<TaskCreateVM>
     {
         public TaskValidator()
         {
-            RuleFor(x => x.Id)
-                .NotNull().WithMessage("mandatory field");
+            RuleFor(x => x.Title)
+                .NotEmpty().WithMessage("mandatory field");
             RuleFor(x => x.Description)
-                .Length(0, 50).WithMessage("Description should be from 0-10 characters");
-            RuleFor(x => x.DateAdded)
-                .LessThan(DateTime.Now).WithMessage("Trying to add past date");
+                .Length(5, 20).WithMessage("Description should be from 5-20 characters");
             RuleFor(x => x.DateDeadline)
-                .GreaterThanOrEqualTo(x => x.DateAdded).WithMessage("Deadline should not be below date added");
-           //RuleFor(x => x.DateClosure)
-           //     .NotNull().WithMessage("DateClosure should be set for closed tasks")
-           //     .When(x => x.Status == StatusList.closed);
-           // RuleFor(x => x.DateClosure)
-           //     .Null().WithMessage("DateClosure is not set when task still open/in progress")
-           //    .When(x => x.Status == StatusList.open || x.Status == StatusList.in_progress);
+                .GreaterThanOrEqualTo(DateTime.Now).WithMessage("Deadline should not be below date added");
+            RuleFor(x => x.DateDeadline)
+                .LessThanOrEqualTo(DateTime.Now.AddDays(7)).WithMessage("High importance task should be completed within 7 days")
+                .When(x => x.Importance == ImportanceList.high);
         }
     }
 }
